# Refactor UOps match logic to UPat instead of dicts