Refactors the `_match` function to use `UPat`s. I currently added a `from_dict` method on `UPat` so that this PR can do the main refactor without also including all the changes to rewrite the patterns from dicts to `UPat`s. If this looks good I'm happy to add those here or in a followup PR (and remove the `from_dict` method).

Speed benchmark (M1 macbook air):

```
# master (236390aa)
***** model forward in  31.49 ms
***** model schedule in  10.47 ms
***** model lower in 2608.46 ms
# this pr
***** model forward in  31.40 ms
***** model schedule in  10.41 ms
***** model lower in 1556.09 ms
```