@Qazalin 2024-05-31
@alecchendev What backend did you benchmark against?

I ran `PROFILE=0 python3 test/external/external_benchmark_schedule.py` on both an M1 mac and tinybox green - This diff causes a regression.
eg. tinybox green
master:
```
***** model forward in  51.64 ms
***** model schedule in  16.26 ms
***** model lower in 2097.72 ms
```
this branch:
```
***** model forward in  52.52 ms
***** model schedule in  16.24 ms
***** model lower in 2141.94 ms
```