@alecchendev 2024-05-31
> What backend did you benchmark against?
> 
> I ran PROFILE=0 python3 test/external/external_benchmark_schedule.py on both an M1 mac and tinybox green - This diff causes a regression.

Huh, interesting, I ran the same command on my M1 macbook (air) to get my results.