@alecchendev 2024-05-31
Oh but actually I'm noticing if I run the test many times I get results closer to yours. Not sure why the first time I had run the test yesterday and today it showed the opposite