@geohot 2024-06-01
Oh this is a much better way to do the bounty!