@geohot 2024-06-01
Merged. @alecchendev the bounty is yours once you merge the second PR.